requirements.txt
alembic==1.4.3
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.0
py-moneyed==0.8.0
pycountry==20.7.3
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
schwifty==2020.9.0
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
Werkzeug==1.0.1
WTForms==2.3.3
